# Optimize macros