Compile duration of a crate that uses deku (median from 5 samples):

Before: 19.1 s
After: 4.0 s

Same crate, with derive replaced with deku_derive:

Before: 4.0 s
After: 3.7 s