@dullbananas 2023-02-19
https://nnethercote.github.io/perf-book/heap-allocations.html

https://nnethercote.github.io/perf-book/iterators.html